キャストが必要

+++ atrank/model.py
@@ -421,7 +421,7 @@ def modal_head_attention(queries,
     att = tf.einsum('bqmi,bkni->bqmkn', Q, K) # (N, T_q, M, T_k, M)
     
     # sqrt(C*kl)で割る
-    ckl = (num_units*keys_length) ** 0.5 # (N)
+    ckl = tf.cast(num_units*keys_length, dtype=tf.float32) ** 0.5 # (N)
     ckl = tf.expand_dims(ckl, 1) # (N, 1)
     ckl = tf.expand_dims(ckl, 2) # (N, 1, 1)
     ckl = tf.expand_dims(ckl, 3) # (N, 1, 1, 1)
